rename files, remove try-excepts, use transitive_libraries fixture
