+++ conan/tools/sbom/cyclonedx.py
