+++ test/functional/sbom/test_cyclonedx.py
@@ -13,18 +13,15 @@ import json
 import os
 from conan.errors import ConanException
 from conan.api.output import ConanOutput
-from conan.tools.sbom.cycloneDX import cyclonedx_1_4
+from conan.tools.sbom.cyclonedx import cyclonedx_1_4
 
 def post_package(conanfile):
-    try:
-        sbom_cyclonedx_1_4 = cyclonedx_1_4(conanfile.subgraph)
-        metadata_folder = conanfile.package_metadata_folder
-        file_name = "cyclonedx_1_4.json"
-        with open(os.path.join(metadata_folder, file_name), 'w') as f:
-            json.dump(sbom_cyclonedx_1_4, f, indent=4)
-        ConanOutput().success(f"CYCLONEDX CREATED - {conanfile.package_metadata_folder}")
-    except Exception as e:
-        ConanException("error generating CYCLONEDX file")
+    sbom_cyclonedx_1_4 = cyclonedx_1_4(conanfile.subgraph)
+    metadata_folder = conanfile.package_metadata_folder
+    file_name = "cyclonedx_1_4.json"
+    with open(os.path.join(metadata_folder, file_name), 'w') as f:
+        json.dump(sbom_cyclonedx_1_4, f, indent=4)
+    ConanOutput().success(f"CYCLONEDX CREATED - {conanfile.package_metadata_folder}")
 """
 
 @pytest.fixture()
@@ -34,15 +31,19 @@ def hook_setup_post_package():
     save(hook_path, sbom_hook_post_package)
     return tc
 
+@pytest.fixture()
+def hook_setup_post_package_tl(transitive_libraries):
+    tc = transitive_libraries
+    hook_path = os.path.join(tc.paths.hooks_path, "hook_sbom.py")
+    save(hook_path, sbom_hook_post_package)
+    return tc
 
-def test_sbom_generation_create(hook_setup_post_package):
-    tc = hook_setup_post_package
-    tc.run("new cmake_lib -d name=dep -d version=1.0")
-    tc.run("export .")
-    tc.run("new cmake_lib -d name=foo -d version=1.0 -d requires=dep/1.0 -f")
-    tc.run("export .")
-    tc.run("new cmake_lib -d name=bar -d version=1.0 -d requires=foo/1.0 -f")
-    # bar -> foo -> dep
+
+
+def test_sbom_generation_create(hook_setup_post_package_tl):
+    tc = hook_setup_post_package_tl
+    tc.run("new cmake_lib -d name=bar -d version=1.0 -d requires=engine/1.0 -f")
+    # bar -> engine/1.0 -> matrix/1.0
     tc.run("create . --build=missing")
     bar_layout = tc.created_layout()
     assert os.path.exists(os.path.join(bar_layout.metadata(), "cyclonedx_1_4.json"))
@@ -71,18 +72,15 @@ import json
 import os
 from conan.errors import ConanException
 from conan.api.output import ConanOutput
-from conan.tools.sbom.cycloneDX import cyclonedx_1_4
+from conan.tools.sbom.cyclonedx import cyclonedx_1_4
 
 def post_generate(conanfile):
-    try:
-        sbom_cyclonedx_1_4 = cyclonedx_1_4(conanfile.subgraph)
-        metadata_folder = conanfile.package_metadata_folder
-        file_name = "cyclonedx_1_4.json"
-        with open(os.path.join(metadata_folder, file_name), 'w') as f:
-            json.dump(sbom_cyclonedx_1_4, f, indent=4)
-        ConanOutput().success(f"CYCLONEDX CREATED - {conanfile.package_metadata_folder}")
-    except Exception as e:
-        ConanException("error generating CYCLONEDX file")
+    sbom_cyclonedx_1_4 = cyclonedx_1_4(conanfile.subgraph)
+    metadata_folder = conanfile.package_metadata_folder
+    file_name = "cyclonedx_1_4.json"
+    with open(os.path.join(metadata_folder, file_name), 'w') as f:
+        json.dump(sbom_cyclonedx_1_4, f, indent=4)
+    ConanOutput().success(f"CYCLONEDX CREATED - {conanfile.package_metadata_folder}")
 """
 
 @pytest.fixture()
